move: Input을 common 파일로 이동
